[specs] Minor re-organization.
This re-organizes some of the macros to try to make the intent
more clear and to separate module-spec and protocol-spec definers
to call into the same underlying macros.

This also brings back support for a module-spec referencing a
protocol-spec and having that work. (This is used in the DUIM tests
and probably others.)